ballon2/common/visitor.go: Drop pos argument from NewCollectable() + tests
The intent was to avoid the redundancy of having to specify the position twice,
once when calling NewCollectable() and another time when creating the underlying
Visitable.

The Visitable interface has been augmented with a Position() method, which in
turn enables simplyfying the Collectable struct, dropping its pos field, which
can now be obtained directly via its underlying Visitable field.

On balloon2/{history,hyper}/prune_test.go, a set of proxy functions have
been added that allows expressing the tests more compactly.

Thanks a bunch to @jllucas for helping me with this one.