Reduce the size of raft snapshots by only saving the last value of each key
Following @aalda's indications, I've ported the method Backup() from the badger
sources to our code, tweaking the iterator options so that only the last value
of each key is saved.

The result of this change is an reduction of 25% in the size of the raft
snapshots (test done by adding 1M elements to qed).


Former-commit-id: 85ded879ea21eedee268646ac602a61cfb4286fd [formerly e12e87da7a2050d6091aaec8f419611e01c709cd]
Former-commit-id: b8c1aa7e6682a4f73a4cee2b18c3a8cb8576a01c