Benchmark: Add IncrementalProof benchmark
The IncrementalProof benchmark is now the default, since it is expected to be
the more common request.

The Membership benchmark can be enabled via the -m (-membership) flag.

The shell script launcher has been renamed from stress-membership to simply
`stress`, since it can now run both IncrementalProof and Membership benchmarks.

Co-authored-by: iknite <enrique@iknite.com>

Former-commit-id: a9deaa81e2cfd4c9e2f2feb52a0f8e635a1ea41f [formerly ffc66cb1337cd9bb964cddd1619a1528e6294b12]
Former-commit-id: 63117af21c354225aaeb116feff169808a72c002