Benchmark: Add IncrementalProof benchmark
The IncrementalProof benchmark is now the default, since it is expected to be
the more common request.

The Membership benchmark can be enabled via the -m (-membership) flag.

The shell script launcher has been renamed from stress-membership to simply
`stress`, since it can now run both IncrementalProof and Membership benchmarks.

Co-authored-by: iknite <enrique@iknite.com>